+++ NEWS.md
@@ -1,3 +1,12 @@
+
+## **_0.5-1_**
+
+**Bug fixes**
+
+* [#57](https://github.com/eblondel/zen4R/issues/57) Fix code cleaning in getDepositionById
+* [#58](https://github.com/eblondel/zen4R/pull/58) Fix getDepositionByConceptDOI method
+* [#60](https://github.com/eblondel/zen4R/issues/60) Fix old file upload API
+
 ## **_0.5_**
 
 **New features**